rename graphics headers .h -> .hpp
