+++ source/graphics/source/graphics_shaders.cpp
@@ -1,4 +1,4 @@
-#include "graphics/graphics_shaders.h"
+#include "graphics/graphics_shaders.hpp"
 
 #include <vector>
 #include <string>
